fixed eslint error at tests for every view an extra file with the test
