+++ package.json
@@ -27,7 +27,6 @@
     "eslint-plugin-react": "7.16.0",
     "eslint-plugin-standard": "4.0.1",
     "flow-bin": "0.49.1",
-    "jest": "^25.1.0",
     "puppeteer": "^2.1.0",
     "react-testing-library": "^8.0.1"
   },
